style. 불필요한 주석 제거 및 코드 삭제 style. any type 변경 remove. 불필요한 파일 삭제
