--- components/ledger/types.ts
@@ -1,9 +1,0 @@
-export interface LedgerData {
-  id: number;
-  memberNumber: string;
-  name: string;
-  phoneNumber: string;
-  receptionist: string;
-  chargeAmount: string;
-  paymentMethod: string;
-}
